+++ core/subbin_sizes.cpp
@@ -100,11 +100,11 @@ SubbinSizes binary(const SubbinSizes &a, const SubbinSizes &b, Op op) {
 }
 
 SubbinSizes operator+(const SubbinSizes &a, const SubbinSizes &b) {
-  return binary(a, b, element::plus_equals);
+  return binary(a, b, element::add_equals);
 }
 
 SubbinSizes operator-(const SubbinSizes &a, const SubbinSizes &b) {
-  return binary(a, b, element::minus_equals);
+  return binary(a, b, element::subtract_equals);
 }
 
 } // namespace scipp::core